chore: move sql package to infrastructure
