+++ internal/exporter/goose/goose_migrations.go
@@ -2,11 +2,13 @@ package goose
 
 import (
 	"fmt"
+
+	"github.com/artarts36/gds"
+
 	"github.com/artarts36/db-exporter/internal/exporter/common"
 	"github.com/artarts36/db-exporter/internal/exporter/migrations"
+	"github.com/artarts36/db-exporter/internal/infrastructure/sql"
 	"github.com/artarts36/db-exporter/internal/shared/goose"
-	"github.com/artarts36/db-exporter/internal/sql"
-	"github.com/artarts36/gds"
 )
 
 func NewMigrationsExporter(
