+++ internal/infrastructure/sql/qb_delete.go
