+++ internal/infrastructure/sql/qb_insert.go
